requirements.txt
asgiref==3.7.2
certifi==2024.6.2
distlib==0.3.7
Django==5.0.2
django-debug-toolbar==4.3.0
djangorestframework==3.14.0
filelock==3.13.1
flake8==7.0.0
mccabe==0.7.0
mysql-connector-python==8.2.0
mysqlclient==2.2.4
platformdirs==4.1.0
protobuf==4.21.12
pycodestyle==2.11.1
pyflakes==3.2.0
pytz==2024.1
setuptools==69.0.2
sqlparse==0.4.4
virtualenv==20.25.0
wheel==0.41.3